+++ src/main/api/level-range-data.ts
@@ -1,0 +1,4 @@
+/**
+ * Can be an array of [min, max], just the [min], or a number representing the minimum value
+ */
+export type LevelRangeData = [number, number?] | number