+++ src/main/api/set-title-data.ts
@@ -1,0 +1,4 @@
+/**
+ * The id, or a pair of ids [primal, praetorian] that are used with the /settitle command to set a badge
+ */
+export type SetTitleData = [number, number?]