+++ src/test/api/zone-data.fixture.ts
@@ -4,5 +4,6 @@ import { defineFixture } from 'efate'
 export const zoneDataFixture = defineFixture<ZoneData>((t) => {
   t.key.as(index => `zone-${index}`)
   t.name.as(index => `Zone ${index}`)
+  t.type.withValue('city')
   t.links?.as(() => [{ title: 'foo', href: 'https://nouri.org' }])
 })