+++ src/test/db/zone.test.ts
@@ -22,6 +22,51 @@ describe(Zone.name, () => {
     })
   })
 
+  describe('type', () => {
+    test(`should be set from the data`, () => {
+      const zone = new Zone(zoneDataFixture.create({ type: 'city' }))
+      expect(zone.type).toEqual('city')
+    })
+  })
+
+  describe('morality', () => {
+    test(`should be set from the data`, () => {
+      const zone = new Zone(zoneDataFixture.create({ morality: ['hero'] }))
+      expect(zone.morality?.hero).toBeTruthy()
+      expect(zone.morality?.vigilante).toBeFalsy()
+    })
+
+    test(`should be optional`, () => {
+      const zone = new Zone(zoneDataFixture.omit('morality').create())
+      expect(zone.morality?.all).toBeTruthy()
+    })
+  })
+
+  describe('levelRange', () => {
+    test(`should be set from the data`, () => {
+      const zone = new Zone(zoneDataFixture.create({ levelRange: [10] }))
+      expect(zone.levelRange?.min).toEqual(10)
+      expect(zone.levelRange?.max).toBeUndefined()
+    })
+
+    test(`should be optional`, () => {
+      const zone = new Zone(zoneDataFixture.omit('levelRange').create())
+      expect(zone.levelRange).toBeUndefined()
+    })
+  })
+
+  describe('notes', () => {
+    test(`should be set from the data`, () => {
+      const zone = new Zone(zoneDataFixture.create({ notes: 'foo' }))
+      expect(zone.notes).toEqual('foo')
+    })
+
+    test(`should be optional`, () => {
+      const zone = new Zone(zoneDataFixture.omit('notes').create())
+      expect(zone.notes).toBeUndefined()
+    })
+  })
+
   describe('links', () => {
     test(`should be set from the data`, () => {
       const zone = new Zone(zoneDataFixture.create({ links: [{ title: 'foo', href: 'bar' }] }))